core: Use `SOLVER_FAVOR` in relaxed mode
This closes the loop on the "developer experience" when composing with
lockfiles. We want to remain flexible and allow adding/removing packages
from the manifest without editing lockfiles. But we still want libsolv
to give preference to lockfile packages when evaluating alternatives.

For example, if the manifest has a request for a `foobar` Provides which
both `foo` and `bar` provide and `foo` is in the lockfile, but not
`bar`, we want libsolv to give priority to `foo`. IOW, the presence of a
package in a lockfile is a signal that we expect the Provides to be met
by that package.

Use libsolv's `SOLVER_FAVOR` for this.