rust: Add license to our internal crates, add CI for it
Start using
https://github.com/Nemo157/cargo-lichking
to validate our crates.

I thought of this when I came across https://gitlab.com/woshilapin/with_tempdir/
and then noticed it was lgplv3, and wondered whether
we e.g. had any incompatible licenses in our dependency graph.
According to this (and looking at the output of `cargo lichking list`)
we're good.  But we need to keep this in mind.