upgrader: Pull using new `timestamp-check-from-rev`
Both libostree and rpm-ostree support downgrade protection. But what
that means is different between the two. For libostree, downgrade
protection means not fetching any commit which is older than what the
current ref is pointing at. For rpm-ostree, it means not *deploying*
any commit which is older than what the current *deployment* is on.

These two are mostly the same most of the time, but can differ. For
example, on a remote ref which has commits A -> B -> C, where the client
is sitting on a deployment from A, downgrade protection should not
prevent the client from upgrading to B even if there is a newer commit
C.

Since there is no hard relation enforced between what the state of the
OSTree ref is locally and deployments (e.g. we fully support users
manually doing `ostree pull`), it doesn't make sense to compare against
the tip of the ref. Instead, use the new `timestamp-check-from-rev`
to tell libostree to compare against our base revision, which is what we
care about.

Closes: https://github.com/coreos/fedora-coreos-tracker/issues/481