rojig-build: Just expand %_isa
I have no idea why I came up with the previous code instead
of just expanding `%_isa`.  The old code blew up for me
trying to use rojig w/silverblue because of the `abi` in:

```
$ rpm -q --provides gawk | grep \(
gawk(abi) = 3.0
gawk(x86-64) = 5.0.1-5.fc31
$
```

We can definitely assume that we have the same value of `_isa`
at rojig build time as for the target.