Move release instructions to RELEASE.md
The instructions for doing a release were outdated. Take the opportunity
to just move it to a dedicated markdown file and breaking it out to make
it easier to follow.

This is one step short of using a release checklist like some of the
other CoreOS projects do. This is a good base to eventually do that if
we want.