ci: request 2G of RAM for compilation
I think the `fork` errors CI is hitting might be due to more stringent
memory limit enforcements in OCP4. Let's be more explicit and actually
request 2G of RAM. We can adjust from there.

(One related thing is CPU requests; it's possible we might need to also
make that explicit and in turn adjust the `_smp_mflags` RPM macro, since
the default is unlikely to be k8s/cgroups-aware.)