transaction: Unlock more reliably at txn finish
Today we unlock in the finalize path, which means that
the unlock can be delayed if something is referencing
the gobject for the transaction, even after it's done.

Worse, we do `g_clear_object (&priv->sysroot)` in the dispose
handler which means that if something calls dispose before
the last unref, we will never unlock!

Fix all this by unlocking when the transaction is done
executing.  At that point we aren't going to do any
further modifications, so there's no reason to hold the lock.
This should be much more reliable.

While we're here also track the locked state via an explicit
boolean rather than keying off whether we happen to have
a sysroot object.