client: Close DBus connection before calling reload
Part of looking at
https://bugzilla.redhat.com/show_bug.cgi?id=1865839
https://github.com/coreos/rpm-ostree/pull/2255/

Basically the transaction can be lifecycle bound to the
DBus connection.  We hold that connection open even
after we get a success response and then call back into the
daemon again to do a reload.

Rework the client code here so that all of the txn
stuff is in a method itself, and move the reload outside
of it.  That way the bus connection will have been closed
reliably by autocleanups before we reload.