core: Disable libdnf plugin loading
The libdnf plugins are not really relevant for us (the only in-tree one
is for RHSM). Let's tell libdnf to not even bother trying to load them
in.

Prompted by the fact that libdnf tries to load the plugins from the
bundled path which doesn't exist:

```
libdnf-WARNING **: 13:49:33.207: Can't read plugin directory
"/usr/libexec/rpm-ostree/lib64/libdnf/plugins/": No such file or
directory
```

In test contexts, this warning causes a `SIGABRT`.