Rework some origin APIs to allocate
Part of the reason we had `RpmOstreeOrigin` in the first place
is so that we can cache values we parse from the keyfile.  However
this is a barrier to oxidizing becuase it requires the Rust side
to hold C-compatible `GString` values and worse `GHashTable`.

This isn't a performance sensitive area and it's really not that
bad to switch over all callsites to use autofree, and that will
allow the Rust side to be pure Rust data structures.