daemon: Drop CreateOSName
This is technically an API break, but that method has never really been
very useful and I doubt it was ever seriously used. There is no wrapper
for it in the CLI client. Let's just delete it and pretend it never
existed.

In the end, I don't think the ability of supporting multiple separate
stateroots is widely used in practice.

Closes: #551