tests: Remove "invalid format specifier" for GDateTime test
Newer GLib (in F33) added `%f`, which made this test start
failing when coreos-assembler was updated to F33.

We can't really rely on GLib never adding more letters, and
it's not like we're going to start adding invalid ones
in our code without seeing it, so just drop the test.