kernel: Suppress dracut running restorecon(setfiles)
Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1911505

OSTree owns SELinux labeling, and we don't want "scripts"
trying to change labels on their own.

This is obviously a bit of a hack, I think we want something more
like a `--disable-selinux` option for dracut or perhaps it should
skip trying to run `restorecon` if it's writing to a temporary
file.

But all those require dracut fixes which we have to wait
potentially a while for, so let's do the quick fix here.