Revert "core: Allow overriding downloaded RPMs target dir"
This reverts commit 6a3e3d807d671a3c437fb60dc4f04ccd90e4adc6.

This isn't as useful in the implementation of a `rpm-ostree compose
extensions` because it doesn't account for locally cached repos where no
downloading happens.

Instead, just let libdnf download the packages to the default location
if it's a remote package and we'll just copy it over to the output dir.