Support xattr translation
This adds support for a treefile "xattr-translation" dictionary, and
will then translate extended attributes when they are committed into the
ostree commit.
This can be used for example to translate "user.ima" (which is a common
attribute used by for example ima-evm-utils) to "security.ima", which
would require root permissions to write to during compose.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>